50% changes made
