+++ .env-template
@@ -1,0 +1,5 @@
+PORT=8080
+USERNAME="root",
+PASSWORD="123456",
+DATABASE="node",
+HOST="localhost"
